move symbol tests to container
